# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Misko API
  version: 0.2.0
  description: >-
    Health, authentication, user management and the laboratory singleton.
    Every error response carries a stable `code` and an English `error` message.
    401 means missing, invalid, expired or revoked credentials; 403 means the
    current role lacks the permission. Roles are read from the database on every
    request, never from the token.
security:
  - bearer: []
paths:
...
  /api/lab:
    get:
      operationId: getLaboratory
      summary: The laboratory singleton (*:read)
      responses:
        '200':
          $ref: '#/components/responses/Laboratory'
        '401':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
    patch:
      operationId: updateLaboratory
      summary: Update laboratory fields (lab:configure); an empty code clears it
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              minProperties: 1
              properties:
                name: {type: string, minLength: 1, maxLength: 120}
                code: {type: string, pattern: '^[A-Za-z0-9._-]{0,32}$'}
                timezone: {type: string, description: IANA time zone name}
      responses:
        '200':
          $ref: '#/components/responses/Laboratory'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
...
  /api/environments:
    get:
      operationId: listEnvironments
      summary: Environments ordered by name (*:read)
      parameters:
        - {name: paradigmKey, in: query, schema: {type: string, pattern: '^[A-Z][A-Z_]{0,39}$'}, description: Only environments of this paradigm}
      responses:
        '200':
          description: Environments
          content:
            application/json:
              schema: {type: object, required: [data], properties: {data: {type: array, items: {$ref: '#/components/schemas/Environment'}}}}
        '400': {$ref: '#/components/responses/Problem'}
        '401': {$ref: '#/components/responses/Problem'}
    post:
      operationId: createEnvironment
      summary: Create an environment with its first revision (apparatus:write)
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/EnvironmentInput'}}}}
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [environment, revision]
                properties:
                  environment: {$ref: '#/components/schemas/Environment'}
                  revision: {$ref: '#/components/schemas/EnvironmentRevision'}
        '400': {$ref: '#/components/responses/Problem'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
  /api/environments/{id}:
    parameters:
      - $ref: '#/components/parameters/ID'
    get:
      operationId: getEnvironment
      summary: One environment (*:read)
      responses:
        '200': {description: Environment, content: {application/json: {schema: {$ref: '#/components/schemas/Environment'}}}}
        '401': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
    patch:
      operationId: updateEnvironment
      summary: Rename an environment or edit its notes (apparatus:write); measurements change only through revisions
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/EnvironmentPatch'}}}}
      responses:
        '200': {description: Updated, content: {application/json: {schema: {$ref: '#/components/schemas/Environment'}}}}
        '400': {$ref: '#/components/responses/Problem'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
  /api/environments/{id}/revisions:
    parameters:
      - $ref: '#/components/parameters/ID'
    get:
      operationId: listEnvironmentRevisions
      summary: Revisions of an environment, oldest first (*:read)
      responses:
        '200':
          description: Revisions
          content:
            application/json:
              schema: {type: object, required: [data], properties: {data: {type: array, items: {$ref: '#/components/schemas/EnvironmentRevision'}}}}
        '401': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
    post:
      operationId: createEnvironmentRevision
      summary: Store new measurements as the next revision (apparatus:write); earlier revisions never change
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/EnvironmentRevisionInput'}}}}
      responses:
        '201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/EnvironmentRevision'}}}}
        '400': {$ref: '#/components/responses/Problem'}
        '401': {$ref: '#/components/responses/Problem'}
        '403': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
  /api/environments/{id}/revisions/{number}:
    parameters:
      - $ref: '#/components/parameters/ID'
      - {name: number, in: path, required: true, schema: {type: integer, minimum: 1}}
    get:
      operationId: getEnvironmentRevision
      summary: One revision by number (*:read)
      responses:
        '200': {description: Revision, content: {application/json: {schema: {$ref: '#/components/schemas/EnvironmentRevision'}}}}
        '401': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
...
components:
  parameters:
    ID: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HS256 token with issuer, audience and expiry; see backend/README.md
    worker:
      type: apiKey
      in: header
      name: Authorization
      description: '"Worker TOKEN" with the token returned when the worker was registered; accepted only on /api/worker routes'
  responses:
    Problem:
      description: Error with a stable code
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Laboratory:
      description: Laboratory
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [name, code, timezone, createdAt, updatedAt]
            properties:
              name: {type: string}
              code: {type: [string, 'null']}
              timezone: {type: string}
              createdAt: {type: string, format: date-time}
              updatedAt: {type: string, format: date-time}
  schemas:
...
    ParameterValues:
      type: object
      additionalProperties: {type: number}
      description: Parameter values by key, in the units of the paradigm manifest
    ConstantKey: {type: string, pattern: '^[A-Z][A-Z_]{0,39}$'}
    Environment:
      type: object
      required: [id, name, paradigmKey, notes, latestRevision, createdAt, updatedAt]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        paradigmKey: {$ref: '#/components/schemas/ConstantKey'}
        notes: {type: [string, 'null']}
        latestRevision: {type: integer, minimum: 1}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    EnvironmentRevision:
      type: object
      required: [id, environmentId, number, paradigmKey, paradigmVersion, apparatus, notes, createdBy, createdAt]
      description: Immutable measurements validated against one paradigm version
      properties:
        id: {type: string, format: uuid}
        environmentId: {type: string, format: uuid}
        number: {type: integer, minimum: 1}
        paradigmKey: {$ref: '#/components/schemas/ConstantKey'}
        paradigmVersion: {type: integer, minimum: 1}
        apparatus: {$ref: '#/components/schemas/ParameterValues'}
        notes: {type: [string, 'null']}
        createdBy: {type: string, format: uuid}
        createdAt: {type: string, format: date-time}
    EnvironmentRevisionInput:
      type: object
      additionalProperties: false
      required: [paradigmVersion, apparatus]
      properties:
        paradigmVersion: {type: integer, minimum: 1}
        apparatus:
          allOf: [{$ref: '#/components/schemas/ParameterValues'}]
          description: Every apparatus parameter of the paradigm version, within range and consistent; no measurement is defaulted
        notes: {type: string, maxLength: 2000}
    EnvironmentInput:
      type: object
      additionalProperties: false
      required: [name, paradigmKey, revision]
      properties:
        name: {type: string, minLength: 1, maxLength: 120}
        paradigmKey: {$ref: '#/components/schemas/ConstantKey'}
        notes: {type: string, maxLength: 2000}
        revision: {$ref: '#/components/schemas/EnvironmentRevisionInput'}
    EnvironmentPatch:
      type: object
      additionalProperties: false
      minProperties: 1
      description: An empty notes value clears the notes
      properties:
        name: {type: string, minLength: 1, maxLength: 120}
        notes: {type: string, maxLength: 2000}
...
    Problem:
      type: object
      additionalProperties: false
      required: [error, code]
      properties:
        error: {type: string}
        code: {type: string}
...